+++ backend/app/tests/conftest.py
@@ -1,0 +1,67 @@
+import pytest
+from fastapi.testclient import TestClient
+from sqlmodel import Session, SQLModel, create_engine
+from sqlmodel.pool import StaticPool
+
+from ..dependencies import get_session
+from ..main import app
+from ..schemas import message as message_schema
+from ..schemas import user as user_schema
+
+# Створення тестової бази даних в пам'яті
+DATABASE_URL = "sqlite:///./test.db"
+engine = create_engine(DATABASE_URL)
+
+
+@pytest.fixture(name="session")
+def session_fixture():
+    engine = create_engine(
+        "sqlite://",
+        connect_args={"check_same_thread": False},
+        poolclass=StaticPool,
+    )
+    SQLModel.metadata.create_all(engine)
+    with Session(engine) as session:
+        yield session
+
+
+@pytest.fixture(name="client")
+def client_fixture(session: Session):
+    def get_session_override():
+        return session
+
+    app.dependency_overrides[get_session] = get_session_override
+    client = TestClient(app)
+    yield client
+    app.dependency_overrides.clear()
+
+
+@pytest.fixture
+def test_user(session: Session):
+    user_data = user_schema.UserCreate(
+        email="existing_user@example.com",
+        password="password123",
+        name="Existing User",
+    )
+    user_data.password = None
+    user = user_schema.User(
+        **user_data.model_dump(), hashed_password="password123"
+    )
+    session.add(user)
+    session.commit()
+    session.refresh(user)
+    return user
+
+
+# @pytest.fixture
+# def test_message(session: Session, test_user: user_schema.User):
+#     message_data = message_schema.MessageCreate(
+#         sender_id=test_user.id,
+#         receiver_id=test_user.id,
+#         content="Hello, World!",
+#     )
+#     message = message_schemaMessage(**message_data.model_dump())
+#     session.add(message)
+#     session.commit()
+#     session.refresh(message)
+#     return message
